[PATRO23-64] Move swagger configuration out of program.cs
