+++ src/api/app/Program.cs
@@ -1,4 +1,5 @@
 using Intive.Patronage2023.Api.Configuration;
+using Intive.Patronage2023.Modules.Example.Api;
 using Intive.Patronage2023.Shared.Abstractions;
 using Intive.Patronage2023.Shared.Abstractions.Commands;
 using Intive.Patronage2023.Shared.Abstractions.Queries;
@@ -8,9 +9,7 @@ using Intive.Patronage2023.Shared.Infrastructure.EventDispachers;
 using Intive.Patronage2023.Shared.Infrastructure.EventHandlers;
 using Intive.Patronage2023.Shared.Infrastructure.Queries.QueryBus;
 using Keycloak.AuthServices.Authentication;
-using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.AspNetCore.HttpLogging;
-using Microsoft.OpenApi.Models;
 
 var builder = WebApplication.CreateBuilder(args);
 
@@ -35,46 +34,7 @@ builder.Services.AddMediatR(cfg =>
 	cfg.RegisterServicesFromAssembly(typeof(Program).Assembly);
 });
 
-builder.Services.AddSwaggerGen(options =>
-{
-	options.SwaggerDoc("v1", new OpenApiInfo
-	{
-		Version = "v1",
-		Title = "Intive Patronage2023 Some Title Api",
-		Description = "An ASP.NET Core Web API for managing bills and more",
-	});
-
-	// Adding authentication to swagger
-	var securityScheme = new OpenApiSecurityScheme
-	{
-		Name = "JWT Authentication",
-		Description = "Enter JWT Bearer token **_only_**",
-		In = ParameterLocation.Header,
-		Type = SecuritySchemeType.Http,
-		Scheme = "bearer", // must be lower case
-		BearerFormat = "JWT",
-		Reference = new OpenApiReference
-		{
-			Id = JwtBearerDefaults.AuthenticationScheme,
-			Type = ReferenceType.SecurityScheme,
-		},
-	};
-	options.AddSecurityDefinition(securityScheme.Reference.Id, securityScheme);
-	options.AddSecurityRequirement(new OpenApiSecurityRequirement
-	{
-		{
-			securityScheme, Array.Empty<string>()
-		},
-	});
-
-	// Searching for all files with ".Api.xml" suffix, which should be api docs,
-	// in build directory and attach them to swagger
-	var xmlFiles = Directory.GetFiles(
-		AppContext.BaseDirectory,
-		"*.Api.xml",
-		SearchOption.TopDirectoryOnly).ToList();
-	xmlFiles.ForEach(xmlFile => options.IncludeXmlComments(xmlFile));
-});
+builder.Services.AddSwagger();
 
 builder.Services.AddControllers();
 
@@ -93,6 +53,7 @@ builder.Services.AddAuthorization();
 
 builder.Services.AddScoped<ICommandBus, CommandBus>();
 builder.Services.AddScoped<IQueryBus, QueryBus>();
+
 var app = builder.Build();
 
 app.UseCors(corsPolicyName);
